requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
google_api_python_client==2.159.0
google_auth_oauthlib==1.2.1
protobuf==5.29.3
python-dotenv==1.0.1
pytz==2022.1
SQLAlchemy==2.0.36
Werkzeug==3.1.3
Gunicorn